Changed home task 6.2
