--- _katisheva/6_2/ru/mera/katisheva/task6_2/States.java
@@ -1,5 +1,0 @@
-package ru.mera.katisheva.task6_2;
-
-public enum States {
-    HIDDEN, NOT_HIDDEN, GONE_HOME
-}
