+++ _katisheva/6_2_updated/ru/mera/katisheva/task6_2/LeadingCheck.java
@@ -1,0 +1,5 @@
+package ru.mera.katisheva.task6_2;
+
+public enum LeadingCheck {
+    NOT_HIDE,HIDE
+}
