+++ _katisheva/6_2_updated/ru/mera/katisheva/task6_2/Person.java
