+++ _katisheva/6_2_updated/ru/mera/katisheva/task6_2/States.java
@@ -1,5 +1,5 @@
 package ru.mera.katisheva.task6_2;
 
 public enum States {
-    HIDDEN, NOT_HIDDEN, GONE_HOME
+    NOT_HIDDEN,HIDDEN,GONE_HOME
 }
